# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
      user:
        $ref: '#/definitions/entity.User'
    type: object
...
  dto.SigninRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  dto.SignupRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
...
  entity.Category:
    properties:
      color:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      icon:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      parent_category_id:
        type: string
      sort_order:
        type: integer
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  entity.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      phone:
        type: string
      public_share_url:
        type: string
      settings: {}
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  response.ErrorBody:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
  response.ErrorEnvelope:
    properties:
      error:
        $ref: '#/definitions/response.ErrorBody'
    type: object
host: api.goen.wisteria.local
info:
  contact: {}
  description: REST API for Goen Project
  title: Goen API
  version: "1.0"
paths:
...
  /auth/me:
    get:
      description: Retrieve the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get Current User
      tags:
      - Auth
  /auth/me/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Upload a new avatar image for the user
      parameters:
      - description: Avatar image file
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Upload Avatar
      tags:
      - Auth
  /auth/me/change-password:
    post:
      consumes:
      - application/json
      description: Change the password for the current user
      parameters:
      - description: Password Change Body
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Change Password
      tags:
      - Auth
  /auth/me/profile:
    patch:
      consumes:
      - application/json
      description: Update user information such as email, phone, displayName, or username
      parameters:
      - description: Profile Update Body
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update User Profile
      tags:
      - Auth
  /auth/me/settings:
    patch:
      consumes:
      - application/json
      description: Update the settings for the currently authenticated user
      parameters:
      - description: Settings mappings
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update User Settings
      tags:
      - Auth
  /auth/refresh:
    post:
      description: Get a new access token using an existing session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Refresh Token
      tags:
      - Auth
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return tokens
      parameters:
      - description: Signin request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SigninRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      summary: User Signin
      tags:
      - Auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Signup request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      summary: User Signup
      tags:
      - Auth
...
  /categories:
    get:
      description: Retrieve all categories, optionally filtered by transaction type
      parameters:
      - description: Transaction type filter (INCOME/EXPENSE)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Category'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List Categories
      tags:
      - Categories
  /categories/{categoryId}:
    get:
      description: Retrieve details of a specific category by ID
      parameters:
      - description: Category ID
        in: path
        name: categoryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get Category
      tags:
      - Categories
...
  /media/{bucket}/{key}:
    get:
      description: Retrieve images via minio reverse proxy
      parameters:
      - description: Bucket Name
        in: path
        name: bucket
        required: true
        type: string
      - description: Object Key
        in: path
        name: key
        required: true
        type: string
      produces:
      - image/*
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      summary: Proxy Media Images
      tags:
      - Public
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"